Moved resources to seperate folder
