+++ mass_api_client/__init__.py
@@ -1,6 +1,1 @@
-from .sample import Sample, DomainSample, IPSample, FileSample, ExecutableBinarySample
-from .report import Report
-from .analysis_request import AnalysisRequest
-from .scheduled_analysis import ScheduledAnalysis
-from .analysis_system import AnalysisSystem
-from .analysis_system_instance import AnalysisSystemInstance
+import mass_api_client.resources
