+++ mass_api_client/resources/__init__.py
@@ -1,0 +1,7 @@
+from .analysis_request import AnalysisRequest
+from .analysis_system import AnalysisSystem
+from .analysis_system_instance import AnalysisSystemInstance
+from .report import Report
+from .sample import DomainSample, IPSample, FileSample, ExecutableBinarySample
+from .scheduled_analysis import ScheduledAnalysis
+
